Account for changes in DB structure
A single line in the ipinfo data can apply to a number of
consecutive prefixes, so to avoid repetition the prefixes
themselves are now in their own table with references back
to a "record" table which contains the information that would
otherwise be duplicated for each prefix that is described by
the same line of data.

As part of this, I've fixed the bug that was meant that multiple
prefixes for a line would overwrite one another (since they had
the same ipmeta "record id").