Complete implementation of lookups for memcache_psql provider (mostly)
There's still a memory leak to fix, because normally records
that are added to a record set after a successful lookup
exist permanently inside a datastructure so freeing a record
set shouldn't free the records themselves for the other providers.

Our records however are created on the fly, so they need to be
freed once the user is finished with the lookup results. I'll
need to add a flag to the ipmeta_record_set_t structure that can
indicate who owns the records and therefore whether they should be
freed when the set is destroyed.