Add "mustfree" flag to ipmeta_record_set_t
If set, the records in the set must be freed whenever the record
set is cleared or destroyed (i.e. the records are not references
to entries in a datastructure, but were created by the provider
when a lookup was requested).

Added an internal API method: ipmeta_record_set_require_free()
that will set the mustfree flag for a record set.

Updated ipmeta_record_set_free() to clear the record set before
freeing it, so that the records themselves are now freed if
necessary.