memcache_psql: fix memory leak and reduce number of memcache lookups
Now that ipmeta_record_set_require_free() exists, the memcache_psql
provider can now use it to avoid leaking memory after every lookup.

Changed the structure of the memcache entries. Instead of 1
entry per prefix plus a "number of entries" counter, we now
store multiple prefixes (as separate lines of CSV) in a single
cache entry. If the entry gets too large, we end the current
entry with a special token to indicate that there is more cached
data in another key, and carry on putting prefix data in the
cache entry for the second key.

This means that most cache hits will just require one memcached
lookup, which should be much more efficient.