Progress toward revamp of load_ipmeta_into_psql_db
Partially complete... more to come.

The idea here is to map all IPInfo prefixes onto the regions that
are already defined in IODA to account for major differences in
the regional divisions seen between IPInfo and our previous
geo-location data.

A region map file is now required as additional input, which
defines which IODA region is the best fit (if any) for a
lat-long that appears in the IPInfo data. This allows us to
stick to using the existing IODA region codes everywhere.

In cases where there is no good region match, or the lat-long
does not appear in the region map file, the region will be set
to the corresponding "Unknown Region" for the country code that
the prefix is assigned to.